+++ src/js/features/routines/state/actions.js
@@ -2,23 +2,27 @@ import * as _types from './types';
 import axios from 'axios';
 import * as _config from '../../../../assets/config.json';
 
-const recievedRoutineSet = (payload = {}) => {
-  const { data } = payload;
-  return { routine: data, type: _types.RECIEVE_ROUTINES_SETS };
+const httpHeader = { header: { 'Content-Type': 'application/json' } };
+const rountinesPath = 'json/object/rountines';
+
+const emitDispatch = (type, actions = {}) => {
+  return { type, ...actions };
 };
 
-const requestRoutineSet = () => {
-  return { type: _types.REQUEST_ROUTINES_SETS };
-};
-
-export const fetchRoutineSet = (routineId = 1) => {
-  const url = _config.apis.heroku + 'basicRoutine?routineId=' + routineId;
-  const header = { header: { 'Content-Type': 'application/json' } };
-  return dispatch => {
-    dispatch(requestRoutineSet());
-    return axios.get(url, header)
-      .then(payload => {
-        return dispatch(recievedRoutineSet(payload));
+export const fetchRoutineList = () => {
+  return (dispatch) => {
+    const url = _config.apis.heroku + rountinesPath;
+    dispatch(emitDispatch(_types.GET_REQUEST_ROUTINE_LIST_START));
+    return axios.get(url, httpHeader)
+      .then((payload) => {
+        const routines = !!payload && !!payload.data && payload.data.payload ? payload.data.payload.routines : [];
+        return dispatch(emitDispatch(_types.GET_REQUEST_ROUTINE_LIST_SUCCESS, { routines }));
+      })
+      .catch((error) => {
+        return dispatch(emitDispatch(_types.GET_REQUEST_ROUTINE_LIST_ERROR, { error }));
+      })
+      .finally(() => {
+        return dispatch(emitDispatch(_types.GET_REQUEST_ROUTINE_LIST_COMPLETED));
       });
   };
 };